requirements.txt
streamlit
weaviate-client
PyPDF2
requests
